Add support for returning field choices.
No code modifications here, just a suggestion to discuss. The [Fields API Standard Fields](http://help.wufoo.com/articles/en_US/SurveyMonkeyArticleType/The-Fields-API#Standardfields) allows for the return of the [choices](http://help.wufoo.com/articles/en_US/SurveyMonkeyArticleType/The-Fields-API#choices) present in a multiple choice question . It'd be beneficial in my particular use case to be able to query these choices and use them dynamically. Apologies for ignorance, here, if this is currently possible via the sub-fields functionality, but it doesn't appear to be.